backend/api/requirements-prod.txt
fastapi==0.115.13
uvicorn==0.34.3
pydantic==2.11.7
pydantic_core==2.33.2
starlette==0.46.2
typing_extensions==4.14.0
annotated-types==0.7.0
anyio==4.9.0
sniffio==1.3.1
h11==0.16.0
idna==3.10
pymongo==4.13.2
dnspython==2.7.0
python-dotenv==1.1.1
